TFS Integration is only for Ultimate

+++ dev-util/idea-ultimate/idea-ultimate-2016.2.0.162.1121.32.ebuild
@@ -52,18 +52,27 @@ src_prepare() {
 	if ! use amd64; then
 		rm -r lib/libpty/linux/x86_64 || die
 		rm -f bin/fsnotifier64 bin/libbreakgen64.so bin/idea64.vmoptions || die
-		rm -r plugins/tfsIntegration/lib/native/linux/x86_64 || die
+		if [[ "${MY_EDITION}" == "IU" ]]
+		then
+			rm -r plugins/tfsIntegration/lib/native/linux/x86_64 || die
+		fi
 	fi
 	if ! use x86; then
 		rm -r lib/libpty/linux/x86 || die
 		rm -f bin/fsnotifier bin/libbreakgen.so bin/idea.vmoptions || die
-		rm -r plugins/tfsIntegration/lib/native/linux/x86 || die
+		if [[ "${MY_EDITION}" == "IU" ]]
+		then
+			rm -r plugins/tfsIntegration/lib/native/linux/x86 || die
+		fi
 	fi
 	rm -f bin/fsnotifier-arm || die
 	rm -r lib/libpty/{win,macosx} || die
 	rm Install-Linux-tar.txt || die
-	rm -r plugins/tfsIntegration/lib/native/{aix,freebsd,hpux,macosx,solaris,win32} || die
-	rm -r plugins/tfsIntegration/lib/native/linux/{arm,ppc} || die
+	if [[ "${MY_EDITION}" == "IU" ]]
+	then
+		rm -r plugins/tfsIntegration/lib/native/{aix,freebsd,hpux,macosx,solaris,win32} || die
+		rm -r plugins/tfsIntegration/lib/native/linux/{arm,ppc} || die
+	fi
 
 	if ! use custom-jdk; then
 		if [[ -d jre ]]; then
